+++ nionswift_plugin/nion_eels_analysis/BackgroundSubtraction.py
@@ -19,8 +19,8 @@ from nion.utils import Registry
 _ = gettext.gettext
 
 
-class EELSBackgroundSubtraction:
-    label = _("EELS Background Subtraction")
+class EELSFitBackground:
+    label = _("EELS Fit Background")
     inputs = {
         "eels_spectrum_data_item": {"label": _("EELS Spectrum")},
         "background_model": {"label": _("Background Model"), "entity_id": "background_model"},
@@ -76,8 +76,52 @@ class EELSBackgroundSubtraction:
         self.computation.set_referenced_xdata("subtracted", self.__subtracted_xdata)
 
 
-class EELSMapping:
-    label = _("EELS Map")
+class EELSSubtractBackground:
+    label = _("EELS Subtract Background")
+    inputs = {
+        "spectrum_image_data_item": {"label": _("EELS Image")},
+        "background_model": {"label": _("Background Model"), "entity_id": "background_model"},
+        "fit_interval_graphics": {"label": _("Fit")},
+        }
+    outputs = {
+        "subtracted": {"label": _("EELS Background Subtracted")},
+    }
+
+    def __init__(self, computation, **kwargs):
+        self.computation = computation
+
+    def execute(self, spectrum_image_data_item: Facade.DataItem, background_model, fit_interval_graphics):
+        try:
+            assert spectrum_image_data_item.xdata.is_datum_1d
+            assert spectrum_image_data_item.xdata.is_navigable
+            assert spectrum_image_data_item.xdata.datum_dimensional_calibrations[0].units == "eV"
+            spectrum_image_xdata = spectrum_image_data_item.xdata
+            # fit_interval_graphics.interval returns normalized coordinates. create calibrated intervals.
+            fit_intervals: typing.List[BackgroundModel.BackgroundInterval] = list()
+            for fit_interval_graphic in fit_interval_graphics:
+                fit_intervals.append(fit_interval_graphic.interval)
+            subtracted_xdata = None
+            if background_model._data_structure.entity:
+                entity_id = background_model._data_structure.entity.entity_type.entity_id
+                for component in Registry.get_components_by_type("background-model"):
+                    if entity_id == component.background_model_id:
+                        integrate_result = component.subtract_background(spectrum_xdata=spectrum_image_xdata, fit_intervals=fit_intervals)
+                        subtracted_xdata = integrate_result["subtracted"]
+            if subtracted_xdata is None:
+                subtracted_xdata = DataAndMetadata.new_data_and_metadata(numpy.zeros(spectrum_image_xdata.navigation_dimension_shape), dimensional_calibrations=spectrum_image_xdata.navigation_dimensional_calibrations)
+            self.__subtracted_xdata = subtracted_xdata
+        except Exception as e:
+            import traceback
+            print(traceback.format_exc())
+            print(e)
+            raise
+
+    def commit(self):
+        self.computation.set_referenced_xdata("subtracted", self.__subtracted_xdata)
+
+
+class EELSMapBackgroundSubtractedSignal:
+    label = _("EELS Map Background Subtracted Signal")
     inputs = {
         "spectrum_image_data_item": {"label": _("EELS Image")},
         "background_model": {"label": _("Background Model"), "entity_id": "background_model"},
@@ -85,7 +129,7 @@ class EELSMapping:
         "signal_interval_graphic": {"label": _("Signal")},
         }
     outputs = {
-        "map": {"label": _("EELS Map")},
+        "map": {"label": _("EELS Signal")},
     }
 
     def __init__(self, computation, **kwargs):
@@ -166,6 +210,38 @@ def subtract_background_from_signal(api: Facade.API_1, window: Facade.DocumentWi
         add_background_subtraction_computation(api, window.library, target_display_item, target_data_item, target_intervals)
 
 
+def subtract_background(api, window):
+    target_display = window.target_display
+    if target_display:
+        target_display_item_data_items = target_display._display_item.data_items
+        for computation in api.library._document_model.computations:
+            if computation.processing_id == "eels.background_subtraction3":
+                if computation.get_input("eels_spectrum_data_item") in target_display_item_data_items and computation.get_output("subtracted") in target_display_item_data_items:
+                    eels_spectrum_data_item = computation.get_input("eels_spectrum_data_item")
+                    eels_spectrum_data_item = api._new_api_object(eels_spectrum_data_item)
+                    fit_interval_graphics = computation.get_input("fit_interval_graphics")
+                    fit_interval_graphics = [api._new_api_object(g) for g in fit_interval_graphics]
+                    background_model = computation.get_input("background_model")
+                    background_model = api._new_api_object(background_model)
+                    source_data_items = api.library._document_model.get_source_data_items(eels_spectrum_data_item._data_item)
+                    if len(source_data_items) == 1 and source_data_items[0].xdata.is_navigable and source_data_items[0].datum_dimension_count == 1:
+                        spectrum_image = api._new_api_object(source_data_items[0])
+                        subtracted = api.library.create_data_item_from_data(numpy.zeros(spectrum_image._data_item.xdata.navigation_dimension_shape), title="{} Background Removed".format(spectrum_image.title))
+                        api.library.create_computation(
+                            "eels.subtract_background",
+                            inputs={
+                                "spectrum_image_data_item": spectrum_image,
+                                "fit_interval_graphics": fit_interval_graphics,
+                                "background_model": background_model,
+                            },
+                            outputs={
+                                "subtracted": subtracted
+                            }
+                        )
+                        window.display_data_item(subtracted)
+                    break
+
+
 def use_signal_for_map(api, window):
     target_display = window.target_display
     target_graphic = target_display.selected_graphics[0] if target_display and len(target_display.selected_graphics) == 1 else None
@@ -202,8 +278,9 @@ def use_signal_for_map(api, window):
                     break
 
 
-Symbolic.register_computation_type("eels.background_subtraction3", EELSBackgroundSubtraction)
-Symbolic.register_computation_type("eels.mapping3", EELSMapping)
+Symbolic.register_computation_type("eels.background_subtraction3", EELSFitBackground)
+Symbolic.register_computation_type("eels.mapping3", EELSMapBackgroundSubtractedSignal)
+Symbolic.register_computation_type("eels.subtract_background", EELSSubtractBackground)
 
 BackgroundModelEntity = Schema.entity("background_model", None, None, {})