+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog (nionswift-eels-analysis)
 0.5.5 (unreleased):
 -------------------
 - Add measure temperature computation.
+- Add menu item to remove/subtract background of spectrum image.
 
 0.5.4 (2021-05-26):
 -------------------